pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = "FastAPI starter template with authentication and migrations"
authors = ["Demetrius Panovitch <demetrius.mp789@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.75.0"
uvicorn = {extras = ["standard"], version = "^0.17.6"}
python-multipart = "^0.0.5"
psycopg2-binary = "^2.9.3"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
alembic = "^1.7.7"
click = "^8.1.2"
sqlmodel = "^0.0.6"
email-validator = "^1.1.3"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
isort = "^5.10.1"
flake8 = "^4.0.1"

[tool.isort]
profile = "black"

[tool.poetry.scripts]
db = "app.scripts:db"
dev = "app.scripts:dev"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
